refactor: move JSON Schema modules under profiles directory
Move src/schemas/ to src/profiles/json-schema/ and rename
schemaRegistry to jsonSchemaRegistry, placing JSON Schema modules
alongside ALPS under a unified profiles/ directory.

- Rename route exports: schemas → jsonSchemaProfiles,
  profiles → alpsProfiles
- Rename route files: schemas.ts → json-schema-profiles.ts,
  profiles.ts → alps-profiles.ts
- Update all imports across routes, middleware, and tests
- Update copilot-instructions.md and DSGEP-005 paths
- Mark DSGEP-005 as Accepted

Closes #570
