+++ apps/api/src/routes/alps-profiles.ts
@@ -4,14 +4,14 @@
 import { alpsRegistry } from '@/profiles/alps/registry.js';
 import { Hono } from 'hono';
 
-export const profiles = new Hono();
+export const alpsProfiles = new Hono();
 
 for (const entry of alpsRegistry) {
   // entry.path is e.g. '/profiles/alps/character/item-v1.json'
   // The router is mounted at '/profiles/alps', so strip the prefix for the route pattern.
   const routePath = entry.path.replace(/^\/profiles\/alps/, '');
 
-  profiles.get(routePath, (c) => {
+  alpsProfiles.get(routePath, (c) => {
     return c.json(entry.profile, 200, {
       'Content-Type': 'application/alps+json',
     });
