+++ apps/api/src/routes/root.test.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 import { app } from '@/app.js';
-import { rootGetResponseV1 } from '@/schemas/root/get-response-v1.js';
+import { rootGetResponseV1 } from '@/profiles/json-schema/root/get-response-v1.js';
 import { Ajv2020 } from 'ajv/dist/2020.js';
 import { beforeEach, describe, expect, it } from 'vitest';
