+++ apps/api/src/routes/root.ts
@@ -3,7 +3,7 @@
 
 import type { NegotiatedContentVariables } from '@/middleware/negotiate-content.js';
 import { negotiateContent } from '@/middleware/negotiate-content.js';
-import { rootGetResponseV1 } from '@/schemas/root/get-response-v1.js';
+import { rootGetResponseV1 } from '@/profiles/json-schema/root/get-response-v1.js';
 import type { Env, Hono as HonoApp } from 'hono';
 import { Hono } from 'hono';
 import { findTargetHandler, isMiddleware } from 'hono/utils/handler';
