+++ apps/api/src/routes/userProfile.test.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 Alex Brandt <alunduil@gmail.com>
 // SPDX-License-Identifier: MIT
 
-import { profileGetResponseV1 } from '@/schemas/profile/get-response-v1.js';
+import { profileGetResponseV1 } from '@/profiles/json-schema/profile/get-response-v1.js';
 import type { UserProfile } from '@genshin/domain';
 import { Ajv2020 } from 'ajv/dist/2020.js';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
