+++ apps/api/src/routes/userProfile.ts
@@ -9,9 +9,9 @@ import type { NegotiatedRequestSchemaVariables } from '@/middleware/negotiate-re
 import { negotiateRequestSchema } from '@/middleware/negotiate-request-schema.js';
 import type { ValidatedRequestBodyVariables } from '@/middleware/validate-request-body.js';
 import { validateRequestBody } from '@/middleware/validate-request-body.js';
+import { profileGetResponseV1 } from '@/profiles/json-schema/profile/get-response-v1.js';
+import { profilePatchRequestV1 } from '@/profiles/json-schema/profile/patch-request-v1.js';
 import { getProfile, updateProfile } from '@/repositories/profile/index.js';
-import { profileGetResponseV1 } from '@/schemas/profile/get-response-v1.js';
-import { profilePatchRequestV1 } from '@/schemas/profile/patch-request-v1.js';
 import { serialiseProfile, type AuthIdentity, type ProfileUpdate } from '@genshin/domain';
 import type { DecodedIdToken } from 'firebase-admin/auth';
 import { Hono } from 'hono';
